ref(cli): mv uninstall -> uninstall mesh command (#4283)
part of #4255
