+++ demo/clean-kubernetes.sh
@@ -7,7 +7,7 @@ source .env
 
 TIMEOUT="${TIMEOUT:-90s}"
 
-bin/osm uninstall -f --mesh-name "$MESH_NAME" --osm-namespace "$K8S_NAMESPACE"
+bin/osm uninstall mesh -f --mesh-name "$MESH_NAME" --osm-namespace "$K8S_NAMESPACE"
 
 for ns in "$BOOKWAREHOUSE_NAMESPACE" "$BOOKBUYER_NAMESPACE" "$BOOKSTORE_NAMESPACE" "$BOOKTHIEF_NAMESPACE" "$K8S_NAMESPACE"; do
     kubectl delete namespace "$ns" --ignore-not-found --wait --timeout="$TIMEOUT" &
